+++ trackstream/tests/helper/__init__.py
@@ -1,0 +1,17 @@
+# -*- coding: utf-8 -*-
+# see LICENSE.rst
+
+"""Test helper functions."""
+
+__all__ = [
+    "BaseClassDependentTests",
+]
+
+##############################################################################
+# IMPORTS
+
+# PROJECT-SPECIFIC
+from .baseclassdependent import BaseClassDependentTests
+
+##############################################################################
+# END