+++ trackstream/tests/helper/tests/__init__.py
@@ -1,0 +1,19 @@
+# -*- coding: utf-8 -*-
+# see LICENSE.rst
+
+"""Test for `~trackstream.tests.helper`."""
+
+__all__ = [
+    # modules
+    "baseclassdependenttests_tests",
+]
+
+
+##############################################################################
+# IMPORTS
+
+# PROJECT-SPECIFIC
+from . import test_baseclassdependenttests as baseclassdependenttests_tests
+
+##############################################################################
+# END